requirements.txt
torch==2.10.0
transformers==4.57.1
requests==2.32.5
numpy==2.3.4
pyyaml==6.0.3
tqdm==4.67.1
matplotlib==3.10.7
pandas==2.3.3
scikit-learn==1.7.2
streamlit==1.48.1
plotly==6.3.0
